Also store reason of deprecation, if any, in the cache
This is done through some mangling of the deprecated name.

Based on the fact that a function name cannot contain ':', we use it as a
delimiter between function name and reason.

So a deprecated name without ':' is just a name, and a deprecated name with at
least a ':' is a name followed by the reason why it's deprecated.

Note that this storage does *not* introduce a new format version ;-)

Fix #41